+++ client/src/components/Navbar.tsx
@@ -30,14 +30,6 @@ export default function Navbar() {
           </div>
           <div className="dropdown">
             <NavLink
-              to="/connect"
-              className={({ isActive }: { isActive: boolean }) => isActive ? 'link active' : 'link'}
-            >
-              Connection Manager
-            </NavLink>
-          </div>
-          <div className="dropdown">
-            <NavLink
               to="/sim"
               className={({ isActive }: { isActive: boolean }) => isActive ? 'link active' : 'link'}
             >
@@ -48,13 +40,25 @@ export default function Navbar() {
                 to="/sim"
                 className={({ isActive }: { isActive: boolean }) => (isActive ? 'link active' : 'link') + ' dropdown-item'}
               >
-                Overview
+                Setup
+              </NavLink>
+              <NavLink
+                to="/run"
+                className={({ isActive }: { isActive: boolean }) => (isActive ? 'link active' : 'link') + ' dropdown-item'}
+              >
+                Run
               </NavLink>
               <NavLink
                 to="/simulation/layout"
                 className={({ isActive }: { isActive: boolean }) => (isActive ? 'link active' : 'link') + ' dropdown-item'}
               >
                 Layout Map
+              </NavLink>
+              <NavLink
+                to="/connect"
+                className={({ isActive }: { isActive: boolean }) => (isActive ? 'link active' : 'link') + ' dropdown-item'}
+              >
+                Server Settings
               </NavLink>
             </div>
           </div>